# openapi.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"description":"Only SUPER ADMIN (level_1) and ADMIN (level_2) can create users","operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully created","content":{"application/json":{"schema":{"example":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4","isActive":true,"createdAt":"2024-01-31T00:00:00.000Z","updatedAt":"2024-01-31T00:00:00.000Z"}}}}},"401":{"description":"Unauthorized - Invalid token"},"403":{"description":"Forbidden - Insufficient permissions"},"409":{"description":"Username already exists"}},"security":[{"JWT-auth":[]}],"summary":"Create a new user","tags":["Users"]},"get":{"description":"Only SUPER ADMIN, ADMIN and MODERATOR can list all users. Supports pagination and search.","operationId":"UsersController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (starts at 1)","schema":{"minimum":1,"default":1,"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page (max 100)","schema":{"minimum":1,"maximum":100,"default":10,"example":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search term to filter users by username, firstName, or lastName","schema":{"example":"john","type":"string"}}],"responses":{"200":{"description":"Paginated list of users","content":{"application/json":{"schema":{"example":{"data":[{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4","isActive":true,"createdAt":"2024-01-31T00:00:00.000Z","updatedAt":"2024-01-31T00:00:00.000Z"}],"meta":{"total":100,"page":1,"limit":10,"totalPages":10,"hasNextPage":true,"hasPreviousPage":false}}}}}},"401":{"description":"Unauthorized - Invalid token"},"403":{"description":"Forbidden - Insufficient permissions"}},"security":[{"JWT-auth":[]}],"summary":"Get all users with pagination","tags":["Users"]}},"/users/{id}":{"get":{"description":"Any authenticated user can view user details","operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"User UUID","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"example":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4","isActive":true,"createdAt":"2024-01-31T00:00:00.000Z","updatedAt":"2024-01-31T00:00:00.000Z"}}}}},"401":{"description":"Unauthorized - Invalid token"},"404":{"description":"User not found"}},"security":[{"JWT-auth":[]}],"summary":"Get user by ID","tags":["Users"]},"patch":{"description":"Only SUPER ADMIN and ADMIN can update users","operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","description":"User UUID","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"User successfully updated","content":{"application/json":{"schema":{"example":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4","isActive":true,"createdAt":"2024-01-31T00:00:00.000Z","updatedAt":"2024-01-31T00:00:00.000Z"}}}}},"401":{"description":"Unauthorized - Invalid token"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"}},"security":[{"JWT-auth":[]}],"summary":"Update user","tags":["Users"]},"delete":{"description":"Only SUPER ADMIN (level_1) can delete users","operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"User UUID","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"responses":{"204":{"description":"User successfully deleted"},"401":{"description":"Unauthorized - Invalid token"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"}},"security":[{"JWT-auth":[]}],"summary":"Delete user","tags":["Users"]}},"/cities":{"post":{"operationId":"CitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCityDto"}}}},"responses":{"201":{"description":""}},"tags":["Cities"]},"get":{"operationId":"CitiesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/cities/{id}":{"get":{"operationId":"CitiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]},"delete":{"operationId":"CitiesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/auth/register":{"post":{"description":"Only SUPER ADMIN (level_1) can register new users","operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered","content":{"application/json":{"schema":{"example":{"access_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","user":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4"}}}}}},"400":{"description":"Bad request - validation error"},"401":{"description":"Unauthorized - Invalid or missing token"},"403":{"description":"Forbidden - Only SUPER ADMIN can register users"},"409":{"description":"Username already exists"}},"security":[{"JWT-auth":[]}],"summary":"Register a new user","tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in","content":{"application/json":{"schema":{"example":{"access_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","user":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4"}}}}}},"401":{"description":"Invalid credentials or user account is inactive"}},"summary":"Login with username and password","tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"User profile retrieved successfully","content":{"application/json":{"schema":{"example":{"id":"123e4567-e89b-12d3-a456-426614174000","username":"johndoe","firstName":"John","lastName":"Doe","role":"level_4","isActive":true}}}}},"401":{"description":"Unauthorized - Invalid or missing token"}},"security":[{"JWT-auth":[]}],"summary":"Get current user profile","tags":["Auth"]}},"/events":{"post":{"description":"Any authenticated user can create events. Level 4 users need approval.","operationId":"EventsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"Event successfully created"},"401":{"description":"Unauthorized"}},"security":[{"JWT-auth":[]}],"summary":"Create a new event","tags":["Events"]},"get":{"description":"Get paginated list of events with optional filters","operationId":"EventsController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (starts at 1)","schema":{"minimum":1,"default":1,"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"minimum":1,"maximum":100,"default":10,"example":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search term to filter users by username, firstName, or lastName","schema":{"example":"john","type":"string"}},{"name":"eventType","required":false,"in":"query","description":"Filtrar por tipo de evento","schema":{"type":"string","enum":["manifestacion","marcha","concentracion","asamblea","otro"]}},{"name":"status","required":false,"in":"query","description":"Filtrar por estado","schema":{"type":"string","enum":["pending","approved","rejected","ongoing","completed","cancelled"]}},{"name":"city","required":false,"in":"query","description":"Filtrar por ciudad","schema":{"example":"La Plata","type":"string"}},{"name":"dateFrom","required":false,"in":"query","description":"Filtrar eventos desde esta fecha","schema":{"example":"2026-01-01","type":"string"}},{"name":"dateTo","required":false,"in":"query","description":"Filtrar eventos hasta esta fecha","schema":{"example":"2026-12-31","type":"string"}}],"responses":{"200":{"description":"Paginated list of events"}},"security":[{"JWT-auth":[]}],"summary":"Get all events with pagination and filters","tags":["Events"]}},"/events/map":{"get":{"description":"Only approved events without pagination for map markers","operationId":"EventsController_findForMap","parameters":[],"responses":{"200":{"description":"List of approved events"}},"security":[{"JWT-auth":[]}],"summary":"Get all approved events for map display","tags":["Events"]}},"/events/pending":{"get":{"description":"Only moderators and admins can see pending events","operationId":"EventsController_findPending","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (starts at 1)","schema":{"minimum":1,"default":1,"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"minimum":1,"maximum":100,"default":10,"example":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search term to filter users by username, firstName, or lastName","schema":{"example":"john","type":"string"}},{"name":"eventType","required":false,"in":"query","description":"Filtrar por tipo de evento","schema":{"type":"string","enum":["manifestacion","marcha","concentracion","asamblea","otro"]}},{"name":"status","required":false,"in":"query","description":"Filtrar por estado","schema":{"type":"string","enum":["pending","approved","rejected","ongoing","completed","cancelled"]}},{"name":"city","required":false,"in":"query","description":"Filtrar por ciudad","schema":{"example":"La Plata","type":"string"}},{"name":"dateFrom","required":false,"in":"query","description":"Filtrar eventos desde esta fecha","schema":{"example":"2026-01-01","type":"string"}},{"name":"dateTo","required":false,"in":"query","description":"Filtrar eventos hasta esta fecha","schema":{"example":"2026-12-31","type":"string"}}],"responses":{"200":{"description":"Paginated list of pending events"}},"security":[{"JWT-auth":[]}],"summary":"Get pending events for moderation","tags":["Events"]}},"/events/{id}":{"get":{"description":"Get detailed information about a specific event","operationId":"EventsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Event UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Event found"},"404":{"description":"Event not found"}},"security":[{"JWT-auth":[]}],"summary":"Get event by ID","tags":["Events"]},"patch":{"description":"Update event. Only owner or moderator+ can edit.","operationId":"EventsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Event UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventDto"}}}},"responses":{"200":{"description":"Event successfully updated"},"403":{"description":"Forbidden"},"404":{"description":"Event not found"}},"security":[{"JWT-auth":[]}],"summary":"Update event","tags":["Events"]},"delete":{"description":"Delete event. Only owner or admin can delete.","operationId":"EventsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Event UUID","schema":{"type":"string"}}],"responses":{"204":{"description":"Event successfully deleted"},"403":{"description":"Forbidden"},"404":{"description":"Event not found"}},"security":[{"JWT-auth":[]}],"summary":"Delete event","tags":["Events"]}},"/events/{id}/updates":{"post":{"description":"Registrar un nuevo seguimiento del evento con datos en tiempo real","operationId":"EventsController_createUpdate","parameters":[{"name":"id","required":true,"in":"path","description":"Event UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventUpdateDto"}}}},"responses":{"201":{"description":"Actualización registrada exitosamente"},"404":{"description":"Evento no encontrado"}},"security":[{"JWT-auth":[]}],"summary":"Agregar actualización a un evento","tags":["Events"]},"get":{"description":"Ver todas las actualizaciones registradas de un evento específico","operationId":"EventsController_getUpdates","parameters":[{"name":"id","required":true,"in":"path","description":"Event UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Historial de actualizaciones"}},"security":[{"JWT-auth":[]}],"summary":"Obtener historial de actualizaciones de un evento","tags":["Events"]}},"/events/{eventId}/updates/{updateId}":{"delete":{"description":"Solo el creador puede eliminar su actualización","operationId":"EventsController_removeUpdate","parameters":[{"name":"updateId","required":true,"in":"path","description":"Update UUID","schema":{"type":"string"}},{"name":"eventId","required":true,"in":"path","description":"Event UUID","schema":{}}],"responses":{"204":{"description":"Actualización eliminada"},"403":{"description":"Sin permisos"},"404":{"description":"Actualización no encontrada"}},"security":[{"JWT-auth":[]}],"summary":"Eliminar una actualización","tags":["Events"]}}},"info":{"title":"Gamma API","description":"API documentation for Gamma application","version":"1.0","contact":{}},"tags":[{"name":"Auth","description":"Authentication endpoints"},{"name":"Users","description":"User management endpoints"}],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Unique username for the user","example":"johndoe","minLength":3},"password":{"type":"string","description":"User password","example":"securePassword123","minLength":6},"firstName":{"type":"string","description":"User first name","example":"John"},"lastName":{"type":"string","description":"User last name","example":"Doe"},"role":{"type":"string","description":"User role","enum":["level_1","level_2","level_3","level_4","level_5"],"example":"level_4","default":"level_4"},"assignedCityIds":{"description":"IDs de ciudades asignadas (solo para usuarios level_4)","example":["uuid-1","uuid-2"],"type":"array","items":{"type":"string"}}},"required":["username","password","firstName","lastName"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Unique username for the user","example":"johndoe","minLength":3},"password":{"type":"string","description":"User password","example":"securePassword123","minLength":6},"firstName":{"type":"string","description":"User first name","example":"John"},"lastName":{"type":"string","description":"User last name","example":"Doe"},"role":{"type":"string","description":"User role","enum":["level_1","level_2","level_3","level_4","level_5"],"example":"level_4","default":"level_4"},"assignedCityIds":{"description":"IDs de ciudades asignadas (solo para usuarios level_4)","example":["uuid-1","uuid-2"],"type":"array","items":{"type":"string"}},"isActive":{"type":"boolean","description":"User active status","example":true}}},"CreateCityDto":{"type":"object","properties":{}},"RegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"Unique username for the user","example":"johndoe","minLength":3},"password":{"type":"string","description":"User password","example":"securePassword123","minLength":6},"firstName":{"type":"string","description":"User first name","example":"John"},"lastName":{"type":"string","description":"User last name","example":"Doe"},"role":{"type":"string","description":"User role","enum":["level_1","level_2","level_3","level_4","level_5"],"example":"level_4","default":"level_4"}},"required":["username","password","firstName","lastName"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"Username for authentication","example":"johndoe","minLength":3},"password":{"type":"string","description":"User password","example":"securePassword123","minLength":6}},"required":["username","password"]},"CreateEventDto":{"type":"object","properties":{"title":{"type":"string","description":"Título del evento","example":"Manifestación por mejores condiciones laborales","maxLength":255},"description":{"type":"string","description":"Descripción detallada del evento","example":"Se convoca a una manifestación pacífica para..."},"eventType":{"type":"string","description":"Tipo de evento","enum":["manifestacion","marcha","concentracion","asamblea","otro"],"example":"manifestacion"},"eventDate":{"type":"string","description":"Fecha y hora del evento","example":"2026-02-15T18:00:00.000Z"},"address":{"type":"string","description":"Dirección completa del evento","example":"Av. de Mayo 1370, Buenos Aires","maxLength":500},"cityId":{"type":"string","description":"ID de la ciudad (partido) de la Provincia de Buenos Aires","example":"uuid-ciudad"},"latitude":{"type":"number","description":"Latitud de la ubicación","example":-34.6037,"minimum":-90,"maximum":90},"longitude":{"type":"number","description":"Longitud de la ubicación","example":-58.3816,"minimum":-180,"maximum":180},"attendeeCount":{"type":"number","description":"Número estimado de asistentes","example":500,"minimum":0}},"required":["title","description","eventType","eventDate","address","cityId","latitude","longitude"]},"UpdateEventDto":{"type":"object","properties":{"title":{"type":"string","description":"Título del evento","example":"Manifestación por mejores condiciones laborales","maxLength":255},"description":{"type":"string","description":"Descripción detallada del evento","example":"Se convoca a una manifestación pacífica para..."},"eventType":{"type":"string","description":"Tipo de evento","enum":["manifestacion","marcha","concentracion","asamblea","otro"],"example":"manifestacion"},"eventDate":{"type":"string","description":"Fecha y hora del evento","example":"2026-02-15T18:00:00.000Z"},"address":{"type":"string","description":"Dirección completa del evento","example":"Av. de Mayo 1370, Buenos Aires","maxLength":500},"cityId":{"type":"string","description":"ID de la ciudad (partido) de la Provincia de Buenos Aires","example":"uuid-ciudad"},"latitude":{"type":"number","description":"Latitud de la ubicación","example":-34.6037,"minimum":-90,"maximum":90},"longitude":{"type":"number","description":"Longitud de la ubicación","example":-58.3816,"minimum":-180,"maximum":180},"attendeeCount":{"type":"number","description":"Número estimado de asistentes","example":500,"minimum":0},"status":{"type":"string","description":"Estado del evento (solo admin/moderador pueden cambiar)","enum":["pending","approved","rejected","ongoing","completed","cancelled"],"example":"approved"}}},"CreateEventUpdateDto":{"type":"object","properties":{"updateTime":{"type":"string","description":"Hora de la actualización","example":"2026-02-15T10:00:00.000Z"},"attendeeCount":{"type":"number","description":"Número de personas presentes","example":20,"minimum":0},"policePresence":{"type":"boolean","description":"Presencia policial","example":true,"default":false},"streetClosure":{"type":"boolean","description":"Corte de calle","example":false,"default":false},"notes":{"type":"string","description":"Notas o comentarios adicionales","example":"Manifestación pacífica, sin incidentes"}},"required":["updateTime","policePresence","streetClosure"]}}}}